Added doors, room managers and item stand.
Doors activated when player collides with the room manager, in future builds the room manager will handle most logic in the rooms.
Item stands are functional and randomising of loot is also working, need to fill up the data table and work on things of that sort.
Got doors opening and closing working, currently on a delay to open back up for debug purpose.