Dodge now keeps momentum
Altered how the super dodge prevention works to prevent it from resetting your velocity after a dodge even if it was just on the ground. it now stops checking for a super-dodge after half a second instead of a full second, making it unlikely the player will have moved far enough to trigger super dodge prevention if they were on the ground the whole time. this may make it vulnerable to certain super dodges, which could require reworking the value again until we get them just right.